Let's try this again ...

+++ src/ENDFtk/resonanceParameters/unresolved/EnergyDependent/JValue/src/generateList.hpp
@@ -7,7 +7,11 @@ generateList( int amux, int amun, int amug, int amuf,
               std::vector< double >&& gg,
               std::vector< double >&& gf ) {
 
-  std::vector< double > list{ 0.0, 0.0, amux, amun, amug, amuf };
+  std::vector< double > list{ 0.0, 0.0,
+                              static_cast< double >( amux ),
+                              static_cast< double >( amun ),
+                              static_cast< double >( amug ),
+                              static_cast< double >( amuf ) };
 
   unsigned int ne = energies.size();
   if ( ( d.size() != ne ) or ( gx.size() != ne ) or
